Move services to library so that library users can use them
