+++ owncloudApp/src/main/java/com/owncloud/android/dependecyinjection/RemoteDataSourceModule.kt
@@ -24,30 +24,30 @@ import com.owncloud.android.data.authentication.datasources.RemoteAuthentication
 import com.owncloud.android.data.authentication.datasources.implementation.OCRemoteAuthenticationDataSource
 import com.owncloud.android.data.capabilities.datasources.RemoteCapabilitiesDataSource
 import com.owncloud.android.data.capabilities.datasources.implementation.OCRemoteCapabilitiesDataSource
-import com.owncloud.android.data.capabilities.network.OCCapabilityService
-import com.owncloud.android.data.server.datasources.RemoteServerInfoDataSource
 import com.owncloud.android.data.files.datasources.RemoteFileDataSource
-import com.owncloud.android.data.server.datasources.implementation.OCRemoteServerInfoDataSource
 import com.owncloud.android.data.files.datasources.implementation.OCRemoteFileDataSource
-import com.owncloud.android.data.server.network.OCServerInfoService
-import com.owncloud.android.data.files.network.OCFileService
-import com.owncloud.android.data.sharing.shares.network.OCShareService
+import com.owncloud.android.data.server.datasources.RemoteServerInfoDataSource
+import com.owncloud.android.data.server.datasources.implementation.OCRemoteServerInfoDataSource
 import com.owncloud.android.data.sharing.sharees.datasources.RemoteShareeDataSource
 import com.owncloud.android.data.sharing.sharees.datasources.implementation.OCRemoteShareeDataSource
-import com.owncloud.android.data.sharing.sharees.network.OCShareeService
 import com.owncloud.android.data.sharing.shares.datasources.RemoteShareDataSource
 import com.owncloud.android.data.sharing.shares.datasources.implementation.OCRemoteShareDataSource
 import com.owncloud.android.data.user.datasources.RemoteUserDataSource
 import com.owncloud.android.data.user.datasources.implementation.OCRemoteUserDataSource
-import com.owncloud.android.data.user.network.OCUserService
 import com.owncloud.android.lib.common.OwnCloudAccount
 import com.owncloud.android.lib.common.SingleSessionManager
-import com.owncloud.android.lib.resources.status.ServerInfoService
-import com.owncloud.android.lib.resources.files.FileService
-import com.owncloud.android.lib.resources.shares.ShareService
-import com.owncloud.android.lib.resources.shares.ShareeService
-import com.owncloud.android.lib.resources.status.CapabilityService
-import com.owncloud.android.lib.resources.users.UserService
+import com.owncloud.android.lib.resources.files.services.FileService
+import com.owncloud.android.lib.resources.files.services.implementation.OCFileService
+import com.owncloud.android.lib.resources.shares.services.ShareService
+import com.owncloud.android.lib.resources.shares.services.ShareeService
+import com.owncloud.android.lib.resources.shares.services.implementation.OCShareService
+import com.owncloud.android.lib.resources.shares.services.implementation.OCShareeService
+import com.owncloud.android.lib.resources.status.services.CapabilityService
+import com.owncloud.android.lib.resources.status.services.ServerInfoService
+import com.owncloud.android.lib.resources.status.services.implementation.OCCapabilityService
+import com.owncloud.android.lib.resources.status.services.implementation.OCServerInfoService
+import com.owncloud.android.lib.resources.users.services.UserService
+import com.owncloud.android.lib.resources.users.services.implementation.OCUserService
 import org.koin.android.ext.koin.androidContext
 import org.koin.dsl.module
 
@@ -56,12 +56,32 @@ val remoteDataSourceModule = module {
     single { OwnCloudAccount(get(), androidContext()) }
     single { SingleSessionManager.getDefaultSingleton().getClientFor(get(), androidContext()) }
 
-    single<CapabilityService> { OCCapabilityService(get()) }
-    single<ShareService> { OCShareService(get()) }
-    single<ShareeService> { OCShareeService(get()) }
-    single<UserService> { OCUserService(get()) }
-    single<FileService> { OCFileService(get()) }
-    single<ServerInfoService>{ OCServerInfoService()}
+    single<CapabilityService> {
+        OCCapabilityService(
+            get()
+        )
+    }
+    single<ShareService> {
+        OCShareService(
+            get()
+        )
+    }
+    single<ShareeService> {
+        OCShareeService(
+            get()
+        )
+    }
+    single<UserService> {
+        OCUserService(
+            get()
+        )
+    }
+    single<FileService> {
+        OCFileService(
+            get()
+        )
+    }
+    single<ServerInfoService> { OCServerInfoService() }
 
     factory<RemoteCapabilitiesDataSource> {
         OCRemoteCapabilitiesDataSource(
