+++ owncloudData/src/main/java/com/owncloud/android/data/sharing/sharees/datasources/implementation/OCRemoteShareeDataSource.kt
@@ -21,7 +21,7 @@ package com.owncloud.android.data.sharing.sharees.datasources.implementation
 
 import com.owncloud.android.data.executeRemoteOperation
 import com.owncloud.android.data.sharing.sharees.datasources.RemoteShareeDataSource
-import com.owncloud.android.lib.resources.shares.ShareeService
+import com.owncloud.android.lib.resources.shares.services.ShareeService
 import org.json.JSONObject
 import java.util.ArrayList
