+++ owncloudData/src/test/java/com/owncloud/android/data/file/datasource/OCRemoteFileDataSourceTest.kt
@@ -20,7 +20,7 @@
 package com.owncloud.android.data.file.datasource
 
 import com.owncloud.android.data.files.datasources.implementation.OCRemoteFileDataSource
-import com.owncloud.android.data.files.network.OCFileService
+import com.owncloud.android.lib.resources.files.services.implementation.OCFileService
 import com.owncloud.android.lib.common.operations.RemoteOperationResult
 import com.owncloud.android.testutil.OC_SERVER_INFO
 import com.owncloud.android.utils.createRemoteOperationResultMock
